requirements.txt
opencv-python==4.8.1.78
Pillow==10.0.1
numpy<2.0
tensorflow>=2.16.0
matplotlib==3.8.0
pyserial==3.5